pyproject.toml
[tool.poetry]
name = "quantum-curator-backend"
version = "0.1.0"
description = ""
authors = ["jakemeout <jacobosity+github@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
SQLAlchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
pydantic = "^2.8.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"
email-validator = "^2.2.0"
bcrypt = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
